refactor: move config/routes/controllers and update server setup
