+++ server/node-server-app/src/config/corsConfig.js
@@ -1,0 +1,19 @@
+const allowedOrigins = [
+    'http://localhost:5173',
+    'https://localhost:5173'
+];
+
+export const corsOptions = {
+    origin: function (origin, callback) {
+        if (!origin) return callback(null, true);
+
+        if (allowedOrigins.includes(origin)) {
+            callback(null, true);
+        } else {
+            callback(new Error(' Not allowed by CORS'));
+        }
+    },
+    credentials: true,
+    methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+    allowedHeaders: ['Content-Type', 'Authorization'] 
+};
