+++ server/node-server-app/src/config/serverConfig.js
@@ -1,0 +1,1 @@
+export const SERVER_PORT = 9006;
