+++ server/node-server-app/src/controller/mainController.js
@@ -1,0 +1,7 @@
+export const getHome = (req, res) => {
+    res.send('Hello World!');
+};
+
+export const getEngineCheckLight = (req, res) => {
+    res.send('Engine Check Light is OFF!');
+};
