+++ server/node-server-app/src/routes/index.js
@@ -1,0 +1,10 @@
+import express from 'express';
+import { getHome, getEngineCheckLight } from '../controller/mainController.js';
+
+const router = express.Router();
+
+router.get('/', getHome);
+
+router.get('/cel', getEngineCheckLight);
+
+export default router;
